% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.plotter.R
\name{create.color.list}
\alias{create.color.list}
\title{Create Color List}
\usage{
create.color.list(config.var, gbl.var, samples)
}
\arguments{
\item{config.var}{config.var}

\item{gbl.var}{gbl.var}

\item{samples}{samples}
}
\description{
Create Color List
}
