% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.plotter.R
\name{create.symbol.list}
\alias{create.symbol.list}
\title{Create Symbol List}
\usage{
create.symbol.list(config.var, split.color.list, gbl.var, samples)
}
\arguments{
\item{config.var}{config.var}

\item{split.color.list}{split.color.list}

\item{gbl.var}{gbl.var}

\item{samples}{samples}
}
\description{
Create Symbol List
}
