% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.plotter.R
\name{draw.legend}
\alias{draw.legend}
\title{Draw Legend}
\usage{
draw.legend(config.var, gbl.var, samples)
}
\arguments{
\item{config.var}{config.var}

\item{gbl.var}{gbl.var}

\item{samples}{samples}
}
\description{
Draw Legend
}
