% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.plotter.R
\name{draw.plot.grid.setup}
\alias{draw.plot.grid.setup}
\title{Draw Plot Grid Setup}
\usage{
draw.plot.grid.setup(config.var, gbl.var)
}
\arguments{
\item{config.var}{config.var}

\item{gbl.var}{gbl.var}
}
\description{
Draw Plot Grid Setup
}
