% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.plotter.R
\name{draw.plot.grid.snp.names}
\alias{draw.plot.grid.snp.names}
\title{Draw SNP labels}
\usage{
draw.plot.grid.snp.names(config.var, gbl.var)
}
\arguments{
\item{config.var}{config.var}

\item{gbl.var}{gbl.var}
}
\description{
Draw SNP labels
}
