% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.plotter.R
\name{draw.plot.ld.plot}
\alias{draw.plot.ld.plot}
\title{Draw Plot LD Plot}
\usage{
draw.plot.ld.plot(config.var, gbl.var)
}
\arguments{
\item{config.var}{config.var}

\item{gbl.var}{gbl.var}
}
\description{
Draw Plot LD Plot
}
