% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.plotter.R
\name{read.config}
\alias{read.config}
\title{Read in configuration file}
\usage{
read.config(config.file, config.var)
}
\arguments{
\item{config.file}{config.file}

\item{config.var}{config.var}
}
\description{
Read in configuration file
}
