% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.plotter.R
\name{retrieve.data}
\alias{retrieve.data}
\title{Retrieve data}
\usage{
retrieve.data(split.snp.file, split.hap.file, config.var, gbl.var)
}
\arguments{
\item{split.snp.file}{split.snp.file}

\item{split.hap.file}{split.hap.file}

\item{config.var}{config.var}

\item{gbl.var}{gbl.var}
}
\description{
Retrieve data
}
