% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.plotter.R
\name{set.image.parameters}
\alias{set.image.parameters}
\title{Set Image Parameters}
\usage{
set.image.parameters(config.var, gbl.var)
}
\arguments{
\item{config.var}{config.var}

\item{gbl.var}{gbl.var}
}
\description{
Set Image Parameters
}
